refactor(spec-546): name the module after its contents, not its birth (t-3)
`git mv agent/handlers/shared.ts agent/handlers/tool-contract.ts`, plus the
31 import specifiers that named it. No symbol changes and no code moves
inside the file — this commit is the rename and the guard edits it forces.

spec-366 extracted this file from tool-specs.ts and named it after the ACT of
extraction rather than its contents. std-51 cl-9: a module that can only be
called `shared` has no interface yet. `git mv` rather than delete-and-create,
so history follows the file and the diff reads as one mechanical specifier
rename — ToolCtx (35 importers), ToolSpec (25), VERBOSE_FIELD (16) and reqCtx
(12) keep their named imports untouched.

Three guard edits, forced and deliberate. Two would otherwise throw ENOENT at
module load rather than fail an assertion:

- guidance-authoring-confined and spec-249-status-overview both did
  readFileSync(join(HANDLERS_DIR, "shared.ts")) and then excluded that name
  from the readdir. The special case existed only to place the file FIRST in
  the concatenated blob so the window's two anchors appeared in order. Both
  anchors live in the same file, which preserves their order itself, so the
  hack is gone rather than repointed. Added .sort() while there: readdirSync
  order is filesystem-dependent and both guards compute a window from
  character positions.
- response-budget-funnel's hardcoded caller list and its endsWith() skip now
  name tool-contract.ts. Its own comment asks for a conscious edit in the same
  change rather than a silent arrival; this is it. t-4 moves them again, to
  doc-state.ts, when formatState leaves.

No assertion deleted, none relaxed, no file skipped.

Proved the confinement window still BITES rather than reading a green suite:
RFS_START=346299, RFS_END=355613, a 9314-character window with both anchors
found and ordered, and COMPLETION_NUDGE / buildSketchBlock( /
relatedIssuesNudge( all present inside it.

That check turned up a pre-existing defect, filed as issue-2 and NOT caused
here: the guard's fourth prose builder, `nudgeForTransition`, does not exist
anywhere in packages/server/src — the only occurrence in the tree is the
guard's own PROSE_BUILDERS entry. Its loop is `while (idx !== -1)`, so an
absent token never enters the body, `offenders` stays empty and the check
passes having looked for nothing. Absent from origin/develop too. Hardening
the guard is out of scope for this Spec.

14 test files touched; 11 change nothing but an import specifier, and only the
3 guards above change anything else. Verified: make typecheck exit 0; real
.d.ts emission exit 0 (--noEmit skips declaration emit); full server suite 707
files / 6552 tests passed, identical counts; the seven source-scanning guards
run explicitly, 45 tests passed.
