+++ packages/server/src/agent/handlers/comments.ts
@@ -43,7 +43,7 @@ import {
   reqCtx,
   resolveRefArg,
   type ToolSpec,
-} from "./shared.js";
+} from "./tool-contract.js";
 
 // ── Moved here from shared.ts by spec-546 t-2: this file is the symbol's only
 // consumer, so it lives with its consumer and is private [per std-51].
