+++ packages/server/src/agent/handlers/decisions.ts
@@ -50,7 +50,7 @@ import {
   reqCtx,
   resolveRefArg,
   type ToolSpec,
-} from "./shared.js";
+} from "./tool-contract.js";
 
 export const decisionsTools: ToolSpec[] = [
   {
