+++ packages/server/src/agent/handlers/docs.ts
@@ -89,7 +89,7 @@ import {
   resolveRefArg,
   type ToolCtx,
   type ToolSpec,
-} from "./shared.js";
+} from "./tool-contract.js";
 
 // spec-448 t-5 (ac-8, ac-37): mutating a doc counts as "seeing" it — advance the
 // caller's doc_views marker to the doc's current version so a self-authored
