+++ packages/server/src/agent/handlers/tool-contract.ts
